[MIRROR] Adds examine message support to the boomerang component (#2627)
* Adds examine message support to the boomerang component (#83391)

Fixes #83353
Fixes https://github.com/NovaSector/NovaSector/issues/2594
Alternative to / Closes #83387 

The other PR is doing some hacky stuff that will still mess up vendor
descriptions - the entire point of the CI test that it's bypassing. This
just moves the message to an additional examine line generated by the
component itself.

![image](https://github.com/tgstation/tgstation/assets/6209658/8fe47bbe-b29a-4ac9-b56a-208324fe9a92)

also made the message green so it stands out a little more in the sea of
blue

* Adds examine message support to the boomerang component

---------

Co-authored-by: Afevis <ShizCalev@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>