[MIRROR] Epinephrine will properly freeze multiple allergic reagents (#2636)
* Epinephrine will properly freeze multiple allergic reagents (#83403)

## About The Pull Request
Thanks to an early return, epinephrine was only freezing the metabolism
of a single allergic reagent at a time; refactored the relevant code to
rectify this.
## Why It's Good For The Game
Fixes https://github.com/tgstation/tgstation/issues/83393
## Changelog
:cl: PapaMichael
fix: Epinephrine will prevent metabolism of all allergic reagents
(instead of just one), if taken with multiple chemicals that one is
allergic to.
/:cl:

* Epinephrine will properly freeze multiple allergic reagents

---------

Co-authored-by: PapaMichael <papa.michael.1996@gmail.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>