[MIRROR] Fixes vendors displaying a broken character when dropping cigarette boxes onto the floor. (#2644)
* Fixes vendors displaying a broken character when dropping cigarette boxes onto the floor. (#83379)

## About The Pull Request

Vendors dispensing cigarette boxes onto the floor would cause a broken
character to appear at the beginning of the line.
Looking into it, this seemed to be due to an attempt at capitalizing the
item names, which interacts with the `\improper` text macro to cause
this issue.
Calling `format_text(...)` beforehand resolves this issue.

Look in the linked issue for more details.
## Why It's Good For The Game

Fixes #82618.
## Changelog
:cl:
spellcheck: When a vendor tells you something dropped onto the floor,
the line no longer starts with a broken character in the case of
cigarette boxes.
/:cl:

* Fixes vendors displaying a broken character when dropping cigarette boxes onto the floor.

---------

Co-authored-by: _0Steven <42909981+00-Steven@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>