[MIRROR] Fully deletes improvised shotgun shells from the game. (#2657)
* Fully deletes improvised shotgun shells from the game. (#83436)

## About The Pull Request
For whatever reason, these were partially deleted but the projectile was
left in. Since the thing that spawns these projectiles is 100% gone from
the game and can't even be spawned by admins there's no reason to keep
this vestigial code.
## Why It's Good For The Game
Why waste lines of code and storage space on something that is now gone?
It's just clutter.
## Changelog
:cl:
no player facing change
/:cl:

* Fully deletes improvised shotgun shells from the game.

---------

Co-authored-by: Da Cool Boss <142358580+DaCoolBoss@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>