[MIRROR] Mindless ghosts can become harvesters (#2658)
* Mindless ghosts can become harvesters (#83405)

## About The Pull Request

Fixes #83374
Fixes #82445
#81665 changed Narsie to transfer minds rather than assigning keys to
constructs.
This only works some amount of the time because observers don't have
minds.
Dying in certain ways (such as: when playing as a construct) also seems
to destroy the mind.
Now we transfer the mind if one exists, otherwise we assign the key
(which creates a mind).

## Changelog

:cl:
fix: Ghosts can more reliably become a harvester.
/:cl:

* Mindless ghosts can become harvesters

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>