[MIRROR] Vines can't pull anchored objects, can pull you off tables (#2689)
* Vines can't pull anchored objects, can pull you off tables (#83410)

## About The Pull Request

Fixes #83072 by preventing the ability from targeting anchored atoms.
Also fixes the logic to only check for obstacles interposed between us
and the target, not ones sharing the tile.

## Why It's Good For The Game

This is how the ability was already supposed to work

## Changelog

:cl:
fix: Venus Man Traps cannot use vines to drag around machines or objects
that are bolted, welded, rooted to, or otherwise are part of the ground
fix: Standing on a table will not prevent you from being grappled by a
Venus Man Trap
/:cl:

* Vines can't pull anchored objects, can pull you off tables

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>