[MIRROR] Adds additional type logging to nested feedback for all printed parts. (#2693)
* Adds additional type logging to nested feedback for all printed parts. (#83466)

## About The Pull Request

What it says on the tin, this adds logging to nested_logging for
autolathes, protolathes, and the mech fabricator.

Also quickly renames a variable name within the mech fabricator because
this thing hasn't been updated in awhile and it was egregious to me.

## Why It's Good For The Game

I wanted to throw together a quick superset chart of what items are
printed most often and how often, and was shocked to learn that it
wasn't logged in this way.

It was recorded somewhat within the silo log, but this is MUCH easier to
use all thing considered, so it's an easy fix.

## Changelog

No front facing changes.

* Adds additional type logging to nested feedback for all printed parts.

---------

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>