[MIRROR] [NO GBP]Fixes labour camp shuttle retrieval message. (#2702)
* [NO GBP]Fixes labour camp shuttle retrieval message. (#83488)

## About The Pull Request
I tried to make the name a proper noun with the /p tag, but /p isn't the
way to do it. I don't know the way to do it, but the name will be
capitalised anyways so lets just remove it.
## Why It's Good For The Game
It's better this way.
## Changelog
:cl:
spellcheck: Fixes labour camp shuttle retrieval message starting with
"/p".
/:cl:

* [NO GBP]Fixes labour camp shuttle retrieval message.

---------

Co-authored-by: Pickle-Coding <58013024+Pickle-Coding@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>