[MIRROR] Chutes no longer play sounds on top of each other. Pipe ejections are a bit quieter too (#2713)
* Chutes no longer play sounds on top of each other. Pipe ejections are a bit quieter too (#83470)

## About The Pull Request

Chutes were playing a sound per item in a stack because the condition
was that the last played sound happened before worldtime+1 which was
always the case, that's in the future. They now only play a single sound
when multiple items enter them at once.
Outlets functioned correctly but the sound for midpipe ejections was a
bit overpowering, so I deadened it a bit(-4 LUFS).
Tested. No observed effect on consecutive insertions.

## Why It's Good For The Game

It is often the case that a conveyor with a large stack of items may be
activated, pushing a large quantity of items into a chute, which would
then fry the eardrums of all those nearby.

## Changelog

:cl:
sound: Disposal chutes will no longer play a sound for each item when
many items enter it at once
/:cl:

* Chutes no longer play sounds on top of each other. Pipe ejections are a bit quieter too

---------

Co-authored-by: Capsandi <38051413+Capsandi@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>