[MIRROR] Halves volume of Gravity Generator (#2745)
* Halves volume of Gravity Generator (#83553)

## About The Pull Request

Halves the volume of sound produced by the gravity generator. This is
done code-side and no change is made to the sound files.

Apparently my tasteful composition of retro futuristic jetsons optimism
juxtaposed with an underlying commentary on the costs associated with
said optimistic attitude(see #83517) was lost on the end users of the
/tg/station project. It has never been as clear to me that my HUGE
FUCKING BRAIN with IQ of more than yours is wasted on the /tg/station
project. Yet, my immense intellect is rivaled only by my nigh saintly
mercy, and thus i have deigned to give you this one line pr.
This is untested code.

## Why It's Good For The Game
People who claim to play the game report that the gravity generator is
still too loud.

## Changelog

:cl:
sound: The volume of the gravity generator has been halved
/:cl:

* Halves volume of Gravity Generator

---------

Co-authored-by: Capsandi <38051413+Capsandi@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>