[MIRROR] DNA vault balancing (#2746)
* DNA vault balancing (#83542)

## About The Pull Request
Reduced amount of animal and plant DNA needed for the DNA vault
objective.
## Why It's Good For The Game
This objective getting completed is incredibly rare, BSA is a fun
station objective and gets completed often, The DNA vault is too tedious
and hopefully this will counteract that a little bit before some kind of
rework comes in to fix the underlying issue of the objective being quite
boring.
## Changelog
:cl: grungussuss
balance: The DNA vault now requires less animal and plant DNA to be
completed
/:cl:

* DNA vault balancing

---------

Co-authored-by: Sadboysuss <96586172+Sadboysuss@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>