[MIRROR] Gives the Lawyer a pet goldfish (#2764)
* Gives the Lawyer a pet goldfish (#83461)

## About The Pull Request

The law office now has a pet, a goldfish named McGill.

## Why It's Good For The Game

I've wanted to integrate aquariums and fishes in general into more
roundstart, since it currently either requires being built and/or
ordered, and thought giving it as a pet to the Law Office would be
pretty cool since the law office doesn't have one and deserves one, and
also just a silly reference.

## Changelog

:cl:
add: The Law office now has a pet goldfish.
/:cl:

* Gives the Lawyer a pet goldfish

* Adds to our maps

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>
Co-authored-by: Mal <13398309+vinylspiders@users.noreply.github.com>