[MIRROR] Fixes runtime with serializing photo metadata (#2774)
* Fixes runtime with serializing photo metadata (#83562)

## About The Pull Request

So while working on an entirely unrelated thing, I noticed runtimes
appearing in my logs after taking a picture.
Specifically a bad index runtime for the 77th line here:

https://github.com/tgstation/tgstation/blob/2f6920105e097479ae1f2fdf2823e2e8fdc59f3b/code/modules/photography/_pictures.dm#L66-L78
And the 199th line here:

https://github.com/tgstation/tgstation/blob/2f6920105e097479ae1f2fdf2823e2e8fdc59f3b/code/datums/datum.dm#L193-L200
This seemed to be caused by the fact `serialize_list(...)` is called
with a null `semvers` parameter, which then causes it to try to add to a
null list.

I asked about this in code general on the main discord, and got told
just putting an empty list in when `serialize_list(...)` is called
should be fine.
This resolves the runtime.
## Why It's Good For The Game

Better to not have two runtimes added to the logs whenever someone takes
a picture.
Fixes #80514.

* Fixes runtime with serializing photo metadata

---------

Co-authored-by: _0Steven <42909981+00-Steven@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>