[MIRROR] print 5 cable coils instead of 1 in lathes (#2775)
* print 5 cable coils instead of 1 in lathes (#83601)

## About The Pull Request
What the title says.
## Why It's Good For The Game
It is incredibly tedious to get a large amount of cables from lathes,
considering you need a lot most of the time (especially in engineering),
it should be printing way more than one piece of cable.
## Changelog
:cl: grungussuss
qol: lathes will now print cables coils in packs of 5
/:cl:

* print 5 cable coils instead of 1 in lathes

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>