[MIRROR] Floor diseases are easier to cure (#2804)
* Floor diseases are easier to cure (#83613)

## About The Pull Request

The floor diseases were supposed to be more of a joke, but it seems to
be a bother for many due to the scarcity of cures.

This PR changes the floor diseases cures to some common chemicals.

## Why It's Good For The Game

They weren't supposed to suffer this much.

## Changelog

:cl:
balance: Floor diseases cures are now common chemicals: Milk, Chlorine,
Space Cleaner
/:cl:

* Floor diseases are easier to cure

---------

Co-authored-by: Andrew <mt.forspam@gmail.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>