[MIRROR] Added the zipzap to the ocean sources (#2807)
* Added the zipzap to the ocean sources (#83645)

## About The Pull Request

In this PR I am just adding the Anxious Zipzap from #83508 to an actual
fishing source, as that was left out of the original PR by mistake. It
compiled properly and loaded as a server.
## Why It's Good For The Game

Adding a way to actually obtain a recently added fish (with the only
other way to interact with the power generation mechanics being an emag)
seems like it would be good overall.
## Changelog
:cl:
fix: Made the anxious zipzap accessible.
/:cl:

* Added the zipzap to the ocean sources

---------

Co-authored-by: jwc1015 <64438862+jwc1015@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>