[MIRROR] Revert "Blade creation ritual returns lost blades" (#2808)
* Revert "Blade creation ritual returns lost blades" (#83594)

Reverts tgstation/tgstation#82592

The point of the limit on blades is to punish heretics for handling them
carelessly, allowing them to undo that for free is lame.

Maybe someone can re-add this as a separate ritual which has its own
requirements, maybe a knowledge requirement.

🆑
del: Heretics can no longer return their stolen blades, that's the point
of the limit
/:cl:

* Revert "Blade creation ritual returns lost blades"

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>