[MIRROR] Adds Soup boiling SFX (#2811)
* Adds Soup boiling SFX (#83649)

## About The Pull Request

In honor of the new sound effects general


https://github.com/tgstation/tgstation/assets/51863163/b2db8ee1-51ef-45a8-9cc8-31594d034ea5

Adds a sound effect that plays when soup is boiling (IE, it only plays
when smoke appears over the soup pot)

It has a sharp falloff and a distance penalty so it will really only be
heard in the kitchen (or though the kitchen's walls), maps with ranges
close to the counter like Metastation notwithstanding.

## Why It's Good For The Game

A touch of atmosphere

## Changelog

:cl: Melbert
sound: Boiling soup now makes a sound.
/:cl:

* Adds Soup boiling SFX

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>