[MIRROR] Bumps rust-g to `3.3.0` (#2813)
* Bumps rust-g to `3.3.0` (#83635)

## About The Pull Request
Release notes: https://github.com/tgstation/rust-g/releases/tag/3.3.0

Mainly bumping in order to use a feature of it in #83138, to optimize
the system a little better.

## Why It's Good For The Game
Faster init when that system will be more widely used is a good thing.
Also keeps us up to date and enables more tools to be available to
coders.

* Bumps rust-g to `3.3.0`

---------

Co-authored-by: GoldenAlpharex <58045821+GoldenAlpharex@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>