[MIRROR] Fix Meta cargo lobby being unconnected to power (#2822)
* Fix Meta cargo lobby being unconnected to power (#83666)

## About The Pull Request
Moves a misplaced wire to be connecting the cargo lobby APC instead of
right next to it. Also adds a missing wire in maintenance from cargo's
maintenance to the vault hallway and another one connecting the APC in
the maintenance above mining.

## Why It's Good For The Game
Allows meta's cargo lobby & mining maintence to actually charge without
someone having to add a wire.

## Changelog
:cl: Goat
fix: meta's cargo lobby and mining maintenance APCs is now connected to
the power line
/:cl:

* Fix Meta cargo lobby being unconnected to power

---------

Co-authored-by: Goat <126099705+Goat-Real@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>