[MIRROR] [METASTATION] Connects the vault APC to the powernet (#2833)
* [METASTATION] Connects the vault APC to the powernet (#83696)

## About The Pull Request
Title. PR #83187 revamped Metastation's cargo bay. Unfortunately, three
of the APCs in the area were disconnected from the powernet: the vault,
the cargo lobby, and cargo maintenance. This fixes that.
Update: The cargo lobby and maintenance issues were fixed by #83666, but
not the vault.
## Why It's Good For The Game
Having areas disconnected from the powernet roundstart is bad
## Changelog
:cl:
fix: Metastation's vault is now connected to the power grid
/:cl:

* [METASTATION] Connects the vault APC to the powernet

---------

Co-authored-by: norsvenska <73006946+norsvenska@users.noreply.github.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>