[MIRROR] nerfs the frog ear rape [MDB IGNORE] (#2840)
* nerfs the frog ear rape (#83691)

## About The Pull Request
These 2 sounds are obnoxiously loud and players have been complaining
about it for a while now.

Old:


https://github.com/tgstation/tgstation/assets/96586172/aa9cd23a-d8a9-4c8e-a7cf-4a075c1e6f8b

New:



https://github.com/tgstation/tgstation/assets/96586172/501caa95-1721-4a16-af92-fa0529febb8c

Keep in mind that is is much louder in game, on video it is quieter
because of the recording software.
## Why It's Good For The Game
Stepping on a mob shouldn't explode your ears outwards like in a
cartoon.
## Changelog
:cl: grungussuss
sound: frog sounds will no longer explode your ears
/:cl:

Co-authored-by: Jacquerel <hnevard@gmail.com>

* nerfs the frog ear rape

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>
Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: NovaBot13 <novasector13@gmail.com>